+++ packages/frontend/features/Panes/CollectionItemsList.tsx
@@ -1,4 +1,8 @@
-import { useCollectionItems } from '@features/Collections';
+import { useCallback } from 'react';
+import {
+  useCollectionItems,
+  useRefreshCollection,
+} from '@features/Collections';
 import { useActiveCollection } from '@features/Preferences';
 import { CollectionList } from '@components/collection/list';
 import { ClientRender } from '@utils';
@@ -20,6 +24,21 @@ export const CollectionItemsList: React.FC<CollectionItemsListProps> = (
     allItems,
   } = useCollectionItems(activeCollection?.id);
 
+  const { mutate: refreshCollection, isLoading: isRefreshing } =
+    useRefreshCollection();
+  const handleRefreshClick = useCallback(() => {
+    if (activeCollection) {
+      const collectionId = activeCollection.id;
+      if (typeof collectionId === 'string' && collectionId !== 'home') {
+        return;
+      }
+
+      refreshCollection({ id: collectionId });
+    } else {
+      console.error(`handleRefreshClick() without active collection`);
+    }
+  }, [activeCollection, refreshCollection]);
+
   return (
     <ClientRender>
       <CollectionList
@@ -33,6 +52,8 @@ export const CollectionItemsList: React.FC<CollectionItemsListProps> = (
         isFetchingMoreItems={isFetchingNextPage}
         onFetchMoreItems={fetchNextPage}
         canFetchMoreItems={hasNextPage}
+        isRefreshing={isRefreshing}
+        onRefresh={handleRefreshClick}
       />
     </ClientRender>
   );